+++ plugin/src/test/scala/com/stratio/cassandra/lucene/partitioning/PartitionerOnTokenTest.scala
@@ -15,6 +15,8 @@
  */
 package com.stratio.cassandra.lucene.partitioning
 
+import java.nio.file.Paths
+
 import com.stratio.cassandra.lucene.IndexException
 import org.junit.runner.RunWith
 import org.scalatest.junit.JUnitRunner
@@ -28,13 +30,13 @@ class PartitionerOnTokenTest extends PartitionerTest {
 
   test("build with zero partitions") {
     intercept[IndexException] {
-      PartitionerOnToken(0, Array("/home/a"))
+      PartitionerOnToken(0, Array("/home/a").map(Paths.get(_)))
     }.getMessage shouldBe "The number of partitions should be strictly positive but found 0"
   }
 
   test("build with negative partitions") {
     intercept[IndexException] {
-      PartitionerOnToken(-1, Array("/home/a"))
+      PartitionerOnToken(-1, Array("/home/a").map(Paths.get(_)))
     }.getMessage shouldBe "The number of partitions should be strictly positive but found -1"
   }
 
@@ -50,17 +52,17 @@ class PartitionerOnTokenTest extends PartitionerTest {
   }
 
   test("num partitions") {
-    PartitionerOnToken(4, Array("/home/a","/home/b","/home/c","/home/d")).numPartitions shouldBe 4
+    PartitionerOnToken(4, Array("/home/a","/home/b","/home/c","/home/d").map(Paths.get(_))).numPartitions shouldBe 4
   }
 
   test("key partition with 1 partition") {
     for (i <- 1 to 10) {
-      PartitionerOnToken(1, Array("/home/a")).partition(key(i)) shouldBe 0
+      PartitionerOnToken(1, Array("/home/a").map(Paths.get(_))).partition(key(i)) shouldBe 0
     }
   }
 
   test("key partition with n partitions") {
-    val partitioner = PartitionerOnToken(10, Array("/home/a","/home/b","/home/c","/home/d","/home/e","/home/f","/home/g","/home/h","/home/i","/home/j"))
+    val partitioner = PartitionerOnToken(10, Array("/home/a","/home/b","/home/c","/home/d","/home/e","/home/f","/home/g","/home/h","/home/i","/home/j").map(Paths.get(_)))
     partitioner.partition(key(0)) shouldBe 8
     partitioner.partition(key(1)) shouldBe 9
     partitioner.partition(key(2)) shouldBe 2
@@ -72,28 +74,28 @@ class PartitionerOnTokenTest extends PartitionerTest {
   }
 
   test("test valid paths set get") {
-    val partitioner = PartitionerOnToken(10, Array("/home/a","/home/b","/home/c","/home/d","/home/e","/home/f","/home/g","/home/h","/home/i","/home/j"))
-    partitioner.pathForPartition(0) shouldBe "/home/a"
-    partitioner.pathForPartition(1) shouldBe "/home/b"
-    partitioner.pathForPartition(2) shouldBe "/home/c"
-    partitioner.pathForPartition(3) shouldBe "/home/d"
-    partitioner.pathForPartition(4) shouldBe "/home/e"
-    partitioner.pathForPartition(5) shouldBe "/home/f"
-    partitioner.pathForPartition(6) shouldBe "/home/g"
-    partitioner.pathForPartition(7) shouldBe "/home/h"
-    partitioner.pathForPartition(8) shouldBe "/home/i"
-    partitioner.pathForPartition(9) shouldBe "/home/j"
+    val partitioner = PartitionerOnToken(10, Array("/home/a","/home/b","/home/c","/home/d","/home/e","/home/f","/home/g","/home/h","/home/i","/home/j").map(Paths.get(_)))
+    partitioner.pathForPartition(0) shouldBe Paths.get("/home/a")
+    partitioner.pathForPartition(1) shouldBe Paths.get("/home/b")
+    partitioner.pathForPartition(2) shouldBe Paths.get("/home/c")
+    partitioner.pathForPartition(3) shouldBe Paths.get("/home/d")
+    partitioner.pathForPartition(4) shouldBe Paths.get("/home/e")
+    partitioner.pathForPartition(5) shouldBe Paths.get("/home/f")
+    partitioner.pathForPartition(6) shouldBe Paths.get("/home/g")
+    partitioner.pathForPartition(7) shouldBe Paths.get("/home/h")
+    partitioner.pathForPartition(8) shouldBe Paths.get("/home/i")
+    partitioner.pathForPartition(9) shouldBe Paths.get("/home/j")
   }
 
   test("testing invalid index in pathForPartition") {
-    val partitioner = PartitionerOnToken(1, Array("/home/a"))
+    val partitioner = PartitionerOnToken(1, Array("/home/a").map(Paths.get(_)))
     intercept [IndexOutOfBoundsException] {
       partitioner.pathForPartition(1)
     }.getMessage shouldBe "partition must be [0,1)"
   }
 
   test("testing invalid index in pathForPartition with -1") {
-    val partitioner = PartitionerOnToken(1, Array("/home/a"))
+    val partitioner = PartitionerOnToken(1, Array("/home/a").map(Paths.get(_)))
     intercept [IndexOutOfBoundsException] {
       partitioner.pathForPartition(-1)
     }.getMessage shouldBe "partition must be [0,1)"